Restructure repo: move generator to src/generator and setup monorepo layout
