+++ src/generator/io.cpp
@@ -2,7 +2,7 @@
 // Created by maciej on 15.04.25.
 //
 
-#include "io.hpp"
+#include "../include/generator/io.hpp"
 
 #include <fstream>
 #include <vector>
